feat(aaa-cert/phase-1-a1): contrast:report script — aaa/aa/sub-aa classification per (text, surface, mode) — 100 of 163 pairs aaa-pass, 0 sub-aa
